requirements.txt
google-genai>=0.1.0
google-adk